Add conic demo
In section 2.2.3 we learned something cool about conics, they can
be uniqely defined by any 5 points.  I added a fun tool to demo this
by allowing the user to pick out any 5 points and we build the conic
curve for that.  Its all hacks and slow as beans, but still fun.